+++ plugins/postgres/dbt/adapters/postgres/impl.py
@@ -1,12 +1,15 @@
+from datetime import datetime
 from dataclasses import dataclass
-from typing import Optional, Set
+from typing import Optional, Set, List, Any
 from dbt.adapters.base.meta import available
 from dbt.adapters.base.impl import AdapterConfig
 from dbt.adapters.sql import SQLAdapter
 from dbt.adapters.postgres import PostgresConnectionManager
 from dbt.adapters.postgres import PostgresColumn
 from dbt.adapters.postgres import PostgresRelation
+from dbt.dataclass_schema import dbtClassMixin, ValidationError
 import dbt.exceptions
+import dbt.utils
 
 
 # note that this isn't an adapter macro, so just a single underscore
@@ -14,8 +17,46 @@ GET_RELATIONS_MACRO_NAME = 'postgres_get_relations'
 
 
 @dataclass
+class PostgresIndexConfig(dbtClassMixin):
+    columns: List[str]
+    unique: bool = False
+    type: Optional[str] = None
+
+    def render(self, relation):
+        # We append the current timestamp to the index name because otherwise
+        # the index will only be created on every other run. See
+        # https://github.com/fishtown-analytics/dbt/issues/1945#issuecomment-576714925
+        # for an explanation.
+        now = datetime.utcnow().isoformat()
+        inputs = (self.columns +
+                  [relation.render(), str(self.unique), str(self.type), now])
+        string = '_'.join(inputs)
+        return dbt.utils.md5(string)
+
+    @classmethod
+    def parse(cls, raw_index) -> Optional['PostgresIndexConfig']:
+        if raw_index is None:
+            return None
+        try:
+            cls.validate(raw_index)
+            return cls.from_dict(raw_index)
+        except ValidationError as exc:
+            msg = dbt.exceptions.validator_error_message(exc)
+            dbt.exceptions.raise_compiler_error(
+                f'Could not parse index config: {msg}'
+            )
+        except TypeError:
+            dbt.exceptions.raise_compiler_error(
+                f'Invalid index config:\n'
+                f'  Got: {raw_index}\n'
+                f'  Expected a dictionary with at minimum a "columns" key'
+            )
+
+
+@dataclass
 class PostgresConfig(AdapterConfig):
     unlogged: Optional[bool] = None
+    indexes: Optional[List[PostgresIndexConfig]] = None
 
 
 class PostgresAdapter(SQLAdapter):
@@ -41,6 +82,10 @@ class PostgresAdapter(SQLAdapter):
             )
         # return an empty string on success so macros can call this
         return ''
+
+    @available
+    def parse_index(self, raw_index: Any) -> Optional[PostgresIndexConfig]:
+        return PostgresIndexConfig.parse(raw_index)
 
     def _link_cached_database_relations(self, schemas: Set[str]):
         """