+++ plugins/postgres/dbt/include/postgres/macros/adapters.sql
@@ -14,6 +14,21 @@
   );
 {%- endmacro %}
 
+{% macro postgres__get_create_index_sql(relation, index_dict) -%}
+  {%- set index_config = adapter.parse_index(index_dict) -%}
+  {%- set comma_separated_columns = ", ".join(index_config.columns) -%}
+  {%- set index_name = index_config.render(relation) -%}
+
+  create {% if index_config.unique -%}
+    unique
+  {%- endif %} index if not exists
+  "{{ index_name }}"
+  on {{ relation }} {% if index_config.type -%}
+    using {{ index_config.type }}
+  {%- endif %}
+  ({{ comma_separated_columns }});
+{%- endmacro %}
+
 {% macro postgres__create_schema(relation) -%}
   {% if relation.database -%}
     {{ adapter.verify_database(relation.database) }}